Moved ideas-related summary data classes to teacher-summary-display/summary-data
