+++ src/assets/wise5/directives/teacher-summary-display/summary-data/IdeasSummaryData.ts
